Fixes incorrect assumption
Not all step based ingest workflows required to have a parent
COLLECTION object. Even when it’s common, there are concrete cases,
like newspapers and books, where the parent object is not a collection.
And more over, it does not require such because the resulting ingested
object are not children of such. This pull request fixes an error
introduced by a recent pull which enabled Contextual CMODEL labels on
step forms and broke book and newspaper drupal tests. Also adds the
ability to enable/disable those contextual messages since the use case
could not be global.